P4ADEV-1525 package refactor
