+++ src/main/java/it/gov/pagopa/pu/workflow/wf/ingestionflow/treasury/opi/TreasuryOpiIngestionWFClient.java
@@ -1,9 +1,9 @@
-package it.gov.pagopa.pu.workflow.wf.ingestionflow.treasuryopi;
+package it.gov.pagopa.pu.workflow.wf.ingestionflow.treasury.opi;
 
 import io.temporal.client.WorkflowClient;
 import it.gov.pagopa.pu.workflow.service.WorkflowService;
-import it.gov.pagopa.pu.workflow.wf.ingestionflow.treasuryopi.wfingestion.TreasuryOpiIngestionWF;
-import it.gov.pagopa.pu.workflow.wf.ingestionflow.treasuryopi.wfingestion.TreasuryOpiIngestionWFImpl;
+import it.gov.pagopa.pu.workflow.wf.ingestionflow.treasury.opi.wfingestion.TreasuryOpiIngestionWF;
+import it.gov.pagopa.pu.workflow.wf.ingestionflow.treasury.opi.wfingestion.TreasuryOpiIngestionWFImpl;
 import org.springframework.stereotype.Service;
 
 import static it.gov.pagopa.pu.workflow.utilities.Utilities.generateWorkflowId;
