+++ src/main/java/it/gov/pagopa/pu/workflow/wf/ingestionflow/treasury/opi/config/TreasuryOpiIngestionWfConfig.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.pu.workflow.wf.ingestionflow.treasuryopi.config;
+package it.gov.pagopa.pu.workflow.wf.ingestionflow.treasury.opi.config;
 
 import io.temporal.workflow.Workflow;
 import it.gov.pagopa.payhub.activities.activity.ingestionflow.UpdateIngestionFlowStatusActivity;
